successfully moved state from healthbar to new parent component, fightpage
