+++ src/App.js
@@ -1,15 +1,13 @@
 import './App.css';
 import ChooseMonster from './components/ChooseMonster/ChooseMonster';
-import HealthBar from './components/HealthBar/HealthBar';
-import TaskBar from './components/TaskBar/TaskBar';
+import FightPage from './components/FightPage/FightPage';
 import Particles from './components/Particles/Particles';
 
 function App() {
   return (
     <>
       <ChooseMonster/>
-      <HealthBar/>
-      <TaskBar/>
+      <FightPage />
       <Particles/>
     </>
   );
