+++ src/components/FightPage/HealthBar/HealthBar.js
@@ -1,40 +1,12 @@
 import React from 'react'
-import { useState } from 'react';
 import { LinearProgress } from '@mui/material';
 import { linearProgressClasses } from '@mui/material/LinearProgress';
 import { styled } from '@mui/material/styles';
 import { Button } from '@mui/material';
 
-function HealthBar() {
-    const [noFirstTask, setFirstTask] = useState(true);
-    const [numOfTasks, setNumOfTasks] = useState(0);
-    const [health, setHealth] = useState(100);
+function HealthBar({health, CompleteTask, AddTask}) {
 
-    function CompleteTask(){
-        if (numOfTasks !== 0 && !noFirstTask){
-            setNumOfTasks(numOfTasks - 1);
-        }
-        Attack((noFirstTask) ? 0 : (health / numOfTasks));
-    }
-
-    // decreases health.
-    function Attack(inc){
-
-        if (health > 0){ 
-            setHealth(health - inc)
-        }
-    }
-
-    function AddTask(){
-
-        if (noFirstTask) {
-            setFirstTask(false);
-        }
-        
-        setNumOfTasks(numOfTasks + 1);
-    }
-
-    function progressBarColour() {
+    function ProgressBarColour() {
         if (health > 70){
             return 'lime'
 
@@ -57,7 +29,7 @@ function HealthBar() {
         border: 'solid 1px black',
         borderRadius: '25px',
         [`& .${linearProgressClasses.bar}`]: {
-            backgroundColor: progressBarColour(),
+            backgroundColor: ProgressBarColour(),
             borderRadius: '25px',
         }
     }));
